Dev3

+++ src/application/services/CoordinationServices.ts
@@ -5,6 +5,7 @@ import Coordination from "@/domain/models/coordination";
 import { ApplicationExceptionName } from "@/application/constants/applicationExceptionName";
 import ApplicationException from "@/application/exceptions/applicationException";
 import ICoordinationRepository from "@/domain/repositories/iCoordinationRepository";
+import IProfessorRepository from "@/domain/repositories/iProfessorRepository";
 import { IUnitOfWork } from "../unitOfWork/iUnitOfWork";
 
 @injectable()
@@ -12,6 +13,8 @@ class CoordinationServices implements ICoordinationServices {
     constructor(
         @inject("CoordinationRepository")
         private readonly coordinationRepository: ICoordinationRepository,
+        @inject("ProfessorRepository")
+        private readonly professorRepository: IProfessorRepository,
         @inject("SequelizeUnitOfWork")
         private readonly unitOfWork: IUnitOfWork
     ) { }
@@ -38,6 +41,11 @@ class CoordinationServices implements ICoordinationServices {
     }
     async DeleteAsync(id: number) {
         return await this.unitOfWork.execute(async (trx) => {
+            if (!(await this.coordinationRepository.FindById(id)))
+                throw new ApplicationException(ApplicationExceptionName.NOT_FOUND, "No coordination was found with the provided id", 404)
+            const professorCount = await this.professorRepository.CountByCoordinationId(id)
+            if (professorCount > 0)
+                throw new ApplicationException(ApplicationExceptionName.INVALID_OPERATION, "Cannot delete a coordination that has professors associated with it", 422)
             return (await this.coordinationRepository.Delete(id, trx))
         })
     }
